requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
typing-extensions==3.7.4.3
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.4
